+++ mod-batteries.ts
@@ -1,0 +1,6 @@
+/** @module batteries */
+
+export { HtmlShell } from "./src/HtmlShell.ts";
+export { byoHtml } from "./src/ssr.ts";
+export { Hat, makeHat } from "./src/miscellaneous/Hat.ts";
+export * from "./src/minifyCss.ts";