+++ mod.ts
@@ -1,7 +1,6 @@
 export * from "@oak/oak";
 export * from "@dklab/oak-routing-ctrl";
-export * from "./src/minifyCss.ts";
-export * from "./src/ssr.ts";
+export { ssr } from "./src/ssr.ts";
 export * from "./src/types.ts";
 export {
   _unstable_useCloudflareWorkersMode,
@@ -13,4 +12,3 @@ export {
 } from "./src/useFullSoak.ts";
 export { getOrigin } from "./src/metastore.ts";
 export { setupDefaultFullsoakLogger } from "./src/utils.ts";
-export { HtmlShell } from "./src/HtmlShell.ts";