+++ src/getClientSideJsForRoute.ts
@@ -1,10 +1,10 @@
 import { FULLSOAK_HTMLSHELL_MAINID } from "./constants.ts";
 import { getGlobalComponentsDirName } from "./metastore.ts";
-import { IS_DEBUG, IS_DEV } from "./getEnv.ts";
+import { IS_DEBUG, PREACT_DEVTOOLS_ENABLED } from "./getEnv.ts";
 
 /**
  * retrieve the initial javascript to be loaded into the html document generated with
- * {@link HtmlShell} via the "static" path `/components/:compName/mount`; if this javascript
+ * {@link HtmlShell} via the "static" path `/:globalComponentsDirName/:compName/mount`; if this javascript
  * content fails to load (for any reason), the SSR document should still be enough for
  * human or bot users (we'll lose any interactivity that's `hydrated` by this javascript though)
  */
@@ -15,7 +15,11 @@ export const getClientSideJsForRoute = (compName: string): string => {
   `;
 
   return `
-  ${IS_DEBUG ? `import "preact/debug";` : IS_DEV && `import "preact/devtools";`}
+  ${
+    IS_DEBUG
+      ? `import "preact/debug";`
+      : PREACT_DEVTOOLS_ENABLED && `import "preact/devtools";`
+  }
   import { hydrate } from "preact";
   import { html } from "htm/preact";
   ${importByRoute}