+++ src/utils.ts
@@ -1,4 +1,5 @@
 import { ConsoleHandler, getLogger, type LevelName, setup } from "@std/log";
+import { normalize } from "@std/path";
 import { getEnv } from "./getEnv.ts";
 import { process } from "./getProcess.ts";
 
@@ -36,7 +37,7 @@ if (!DENO_DIR) {
     : `${HOME}/.cache`; // @TODO consider adding support for other systems?
 }
 
-export { DENO_DIR };
+export { DENO_DIR, OS };
 
 /**
  * read a file to string, supporting both file system path and network path
@@ -82,3 +83,11 @@ export const setupDefaultFullsoakLogger = (
     },
   });
 };
+
+/**
+ * handles convering `/` to `\` on Windows
+ * while being no-op for other platforms
+ */
+export const getPlatformAwarePath = OS === "windows"
+  ? (path: string): string => normalize(path)
+  : (path: string): string => path;